refactor(containers): add abstract web container
