+++ src/tomodachi_testcontainers/containers/wiremock.py
@@ -7,11 +7,11 @@ from typing import Any
 
 from testcontainers.core.waiting_utils import wait_for_logs
 
-from tomodachi_testcontainers.containers.common import DockerContainer
+from tomodachi_testcontainers.containers.web import WebContainer
 from tomodachi_testcontainers.utils import copy_folder_to_container
 
 
-class WireMockContainer(DockerContainer):
+class WireMockContainer(WebContainer):
     MAPPINGS_DIR: Path = Path("/home/wiremock/mappings/")
     FILES_DIR: Path = Path("/home/wiremock/__files/")
 
@@ -26,13 +26,10 @@ class WireMockContainer(DockerContainer):
         verbose: bool = False,
         **kwargs: Any,
     ) -> None:
-        super().__init__(image, **kwargs)
+        super().__init__(image, internal_port=internal_port, edge_port=edge_port, **kwargs)
+
         self.mapping_stubs = mapping_stubs
         self.mapping_files = mapping_files
-
-        self.internal_port = internal_port
-        self.edge_port = edge_port
-        self.with_bind_ports(self.internal_port, self.edge_port)
 
         if verbose:
             self.with_command("--verbose")
@@ -40,16 +37,8 @@ class WireMockContainer(DockerContainer):
     def log_message_on_container_start(self) -> str:
         return f"Wiremock admin: http://localhost:{self.edge_port}/__admin"
 
-    def get_internal_url(self) -> str:
-        ip = self.get_container_internal_ip()
-        return f"http://{ip}:{self.internal_port}"
-
-    def get_external_url(self) -> str:
-        host = self.get_container_host_ip()
-        return f"http://{host}:{self.edge_port}"
-
-    def start(self, timeout: float = 10.0) -> "WireMockContainer":
-        super().start()
+    def start(self, timeout: float = 10.0, interval: float = 0.5, status_code: int = 200) -> "WireMockContainer":
+        super().start(timeout=timeout, interval=interval, status_code=status_code)
         wait_for_logs(self, "port:", timeout=timeout)
         self.copy_mappings_to_container()
         self.copy_mapping_files_to_container()
